Develop

+++ controllers/cards.js
@@ -6,25 +6,33 @@ module.exports.createCard = (req, res) => {
 
   Card.create({ name, link, owner })
     .then((card) => res.send({ data: card }))
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch((err) => res.status(500).send({ message: err.message }));
 };
 
 module.exports.getCards = (req, res) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch((err) => res.status(500).send({ message: err.message }));
 };
 
 module.exports.deleteCardById = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .then((card) => {
-      if (card) {
-        res.send({ data: card });
-      } else {
+      if (card) { // карточка найдена
+        if (card.owner.toString() === req.user._id) { // пользователь создатель карточки
+          Card.findByIdAndRemove(req.params.cardId) // удаляем карточку
+            .then((deletedCard) => {
+              res.send({ data: deletedCard });
+            })
+            .catch((err) => res.status(500).send({ message: err.message }));
+        } else { // пользователь не создатель карточки
+          res.status(401).send({ message: 'У вас нет прав для удаления этой карточки' });
+        }
+      } else { // карточка не найдена
         res.status(404).send({ message: 'Карточка не найдена' });
       }
     })
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch((err) => res.status(500).send({ message: err.message }));
 };
 
 module.exports.likeCard = (req, res) => {
@@ -36,7 +44,7 @@ module.exports.likeCard = (req, res) => {
         res.status(404).send({ message: 'Карточка не найдена' });
       }
     })
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch((err) => res.status(500).send({ message: err.message }));
 };
 
 module.exports.dislikeCard = (req, res) => {
@@ -48,5 +56,5 @@ module.exports.dislikeCard = (req, res) => {
         res.status(404).send({ message: 'Карточка не найдена' });
       }
     })
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch((err) => res.status(500).send({ message: err.message }));
 };
